Develop

+++ .github/workflows/create_package.py
@@ -1,14 +1,24 @@
 import os
 import sys
+import shutil
+from pathlib import Path
+import subprocess
 
 def create_package_tree(package_name, autor="Seu Nome"):
     # Crie o diretório do pacote
     os.makedirs(package_name, exist_ok=True)
     os.makedirs(f"{package_name}/{package_name}", exist_ok=True)  # Diretório para o código do pacote
 
-    # Crie o arquivo __init__.py dentro do diretório do pacote
-    with open(f"{package_name}/{package_name}/__init__.py", 'w') as f:
-        f.write(f"# {package_name} - Pacote Python\n")
+    gitignore_path = ".gitignore"
+    setup_content_path = f"{package_name}/setup.py"
+    readme_path = f"{package_name}/README.md"
+    license_path = f"{package_name}/LICENSE"
+
+    if not os.path.exists('./__init__.py'):
+        print("Arquivo não existe")
+        # Crie o arquivo __init__.py dentro do diretório do pacote
+        with open(f"{package_name}/{package_name}/__init__.py", 'w') as f:
+            f.write(f"# {package_name} - Pacote Python\n")
 
     # Crie o arquivo setup.py
     setup_content = f"""from setuptools import setup, find_packages
@@ -31,15 +41,19 @@ setup(
     ],
 )
 """
-    with open(f"{package_name}/setup.py", 'w') as f:
-        f.write(setup_content)
+
+    if not os.path.exists(setup_content_path):
+        with open(setup_content_path, 'w') as f:
+            f.write(setup_content)
 
     # Crie o arquivo README.md
     readme_content = f"""# {package_name}
 
 Este é o pacote `{package_name}`. Adicione aqui uma descrição mais detalhada do seu pacote.
 """
-    with open(f"{package_name}/README.md", 'w') as f:
+    if not os.path.exists(readme_path):
+        print("Arquivo não existe")
+    with open(readme_path, 'w') as f:
         f.write(readme_content)
 
     # Crie o arquivo LICENSE (pode adicionar uma licença como MIT ou qualquer outra)
@@ -65,7 +79,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 """
-    with open(f"{package_name}/LICENSE", 'w') as f:
+    with open(license_path, 'w') as f:
         f.write(license_content)
 
     # Arquivo .gitignore (se necessário)
@@ -75,18 +89,61 @@ dist/
 build/
 *.egg-info
 """
-    with open(f"{package_name}/.gitignore", 'w') as f:
-        f.write(gitignore_content)
+    if not os.path.exists(gitignore_path):
+        print("Arquivo não existe")
+        with open(gitignore_path, 'w') as f:
+            f.write(gitignore_content)
 
     print(f"Estrutura do pacote '{package_name}' criada com sucesso!")
+
+def create_whl_package(package_name):
+    # Verifique se o diretório do projeto existe
+    if not os.path.isdir(package_name):
+        print(f"Erro: O diretório {package_name} não existe!")
+        return
+
+    # Navegue até o diretório do projeto
+    os.chdir(package_name)
+
+    # Verifique se o arquivo setup.py está presente
+    if not os.path.isfile("setup.py"):
+        print("Erro: O arquivo 'setup.py' não encontrado no diretório.")
+        return
+
+    print(f"Gerando o pacote .whl para o projeto '{package_name}'...")
+
+    # Execute o comando para gerar o pacote .whl usando setuptools e wheel
+    try:
+        subprocess.check_call([sys.executable, "setup.py", "bdist_wheel"])
+        print("Pacote .whl gerado com sucesso!")
+    except subprocess.CalledProcessError as e:
+        print(f"Erro ao gerar o pacote: {e}")
 
 if __name__ == "__main__":
     # Verifica se o nome do pacote foi fornecido como argumento
     if len(sys.argv) < 2:
         print("Uso: python create_package_tree.py <package_name> [autor]")
+        sys.exit()
     else:
         package_name = sys.argv[1]
         # Verifica se o autor foi fornecido como o segundo argumento
         autor = sys.argv[2] if len(sys.argv) > 2 else "Seu Nome"
         create_package_tree(package_name, autor)
 
+        # Move arquivos
+        path = Path("./")
+        destination = f"{package_name}/{package_name}/"
+        for file in path.iterdir():
+            if (
+                str(file).startswith(".") or
+                str(file) == "__pycache__" or
+                str(file) == package_name
+            ):
+                pass
+            else:
+                origin = f"./{file}"
+                shutil.move(str(origin), str(destination))
+    try:
+        create_whl_package(package_name)
+    except Exception as e:
+        print(f"Erro ao empacotar script: {e}")
